commit 3: rename, fix typo
