+++ client/pages/to-do/to-do.html
@@ -3,10 +3,10 @@
 
 <main>
     <div class="container">
-        <h1>Welcome User! To Your Personal To Do</h1>
+        <h1>Welcome User!</h1>
         <button id="add-to-do-button" class="btn btn-outline-success">Add To Do</button>
         <!-- LISTING TO DO -->
-        <h2 class="mt-4">Here is Your To Do List:</h2>
+        <h3 class="mt-4">Here is Your To Do List:</h3>
         <div class="row mb-3">
             <div class="col-md-9 themed-grid-col">
                 <div class="pb-3"><h5>#1 To Do Title</h5></div>
