+++ server/app.js
@@ -6,7 +6,7 @@ const { errorHandler } = require("./middlewares/errorHandler")
 const app = express()
 const PORT = parseInt(process.env.PORT) || 3000
 
-const router = require("./routes")
+const router = require("./routes/route")
 
 app.use(express.urlencoded({ extended: false }))
 app.use(express.json())
